feat(billing): usage metering becomes real — a daily sweep derives EmailsDelivered (email_queue terminal states UNIONed with per-attempt delivery_log successes, deduplicated per email under one idempotency marker so a day can never double-record), DedicatedIpHours/StorageGbHours from the platform's own tables, and an ApiCalls batch ingest endpoint (no request-log table exists to sweep — handoff documented), while a reconciliation job compares reserved vs delivered per tenant-day into report rows that flag systematic over-reservation and distinguish bounces from leaks without ever auto-refunding; wallets widen to BIGINT end-to-end (the last int4 cast in a money path removed and gated crate-wide), PAYG millicents round half-up, VAT map-misses become a logged configurable policy instead of a silent Estonian default, dunning transitions emit real events, proration preview serves the deployed config, and an 18-test money-invariants module proves rounding against exact i128 rationals plus a 200-case wallet-conservation property (migration 104)

+++ services/mail-server/crates/billing-service/src/credit_notes.rs
@@ -211,14 +211,14 @@ pub async fn create_credit_note(
                 r#"
                 WITH credit_wallet AS (
                     UPDATE wallets
-                    SET balance = balance + $2::int4,
+                    SET balance = balance + $2::int8,
                         updated_at = NOW()
                     WHERE tenant_id = $1
                     RETURNING id, balance
                 )
                 INSERT INTO wallet_transactions
                     (wallet_id, tenant_id, type, amount, balance_after, description, reference, created_at)
-                SELECT id, $1, 'credit', $2::int4, balance, $3, $4, $5
+                SELECT id, $1, 'credit', $2::int8, balance, $3, $4, $5
                 FROM credit_wallet
                 "#,
             )
